Move call_to_py_function into generators folder
