+++ kybra/compiler/kybra_generate/src/py_ast/kybra_types/kybra_stmt/system_methods/mod.rs
@@ -1,9 +1,8 @@
+use cdk_framework::{nodes::ActFnParam, ToActDataType};
 use proc_macro2::TokenStream;
-use quote::{format_ident, quote};
 
 use super::KybraStmt;
-use crate::{generators::tuple, py_ast::kybra_types::KybraExpr};
-use cdk_framework::{nodes::ActFnParam, ToActDataType};
+use crate::{generators::canister_methods, py_ast::kybra_types::KybraExpr};
 
 mod errors;
 
@@ -40,43 +39,6 @@ impl KybraStmt<'_> {
     }
 
     pub fn generate_call_to_py_function(&self) -> TokenStream {
-        match &self.stmt_kind.node {
-            rustpython_parser::ast::StmtKind::FunctionDef { .. } => {
-                let function_name = self.get_function_name();
-                let act_params = self.build_params();
-
-                let param_conversions = act_params
-                    .iter()
-                    .map(|act_fn_param| {
-                        let name = format_ident!("{}", act_fn_param.prefixed_name());
-                        quote! {
-                            #name.try_into_vm_value(vm).unwrap()
-                        }
-                    })
-                    .collect();
-                let params = tuple::generate_tuple(&param_conversions);
-
-                quote! {
-                    let _kybra_interpreter = _KYBRA_INTERPRETER_OPTION.as_mut().unwrap();
-                    let _kybra_scope = _KYBRA_SCOPE_OPTION.as_mut().unwrap();
-
-                    _kybra_interpreter.enter(|vm| {
-                        let method_py_object_ref = _kybra_unwrap_rust_python_result(_kybra_scope.globals.get_item(#function_name, vm), vm);
-
-                        let result_py_object_ref = vm.invoke(&method_py_object_ref, #params);
-
-                        match result_py_object_ref {
-                            Ok(py_object_ref) => py_object_ref.try_from_vm_value(vm).unwrap(),
-                            Err(err) => {
-                                let err_string: String = err.to_pyobject(vm).repr(vm).unwrap().to_string();
-
-                                panic!("{}", err_string);
-                            }
-                        }
-                    });
-                }
-            }
-            _ => panic!("{}", self.not_a_function_def_error()),
-        }
+        canister_methods::generate_call_to_py_function(self)
     }
 }
